Add remaining project files, BackOffice changes, and credential cleanup
- Add new model files: CreditLimitCheck, DTOPeriodeDates, LoginInfo
- Add appsettings.json with placeholder credentials
- Add BackOffice supplier module (SupplierController, SupplierRepository)
- Add ExcelHelper utilities and PembelianHelper
- Replace hardcoded credentials in BackOffice DatabaseHelper with placeholders
- Add .bak to .gitignore
- Include all pending BackOffice and Penjualan modifications

+++ BackOffice/Model/DTOFakturPembelianDetail.cs
@@ -2,10 +2,10 @@
 {
     public class DTOFakturPembelianDetail
     {
-        public double PURCHASE_ID { get; set; }
+        public int PURCHASE_ID { get; set; }
         public string NO_TRANSAKSI { get; set; }
         public int BARIS { get; set; }
-        public Double PRODUCT_ID { get; set; }
+        public int PRODUCT_ID { get; set; }
         public string KODE_BARANG { get; set; }
         public string NAMA_BARANG { get; set; }
         public string SATUAN { get; set; }
